+++ vitrina/orgs/urls.py
@@ -29,8 +29,11 @@ from vitrina.orgs.views import (
     ContactDeleteView,
     create_remote_organization,
     check_organization,
-    OrganizationAgreementListView,
-    OrganizationAgreementDetailView,
+    OrganizationBasedAgreementListView,
+    OrganizationBasedAgreementDetailView,
+    OrganizationBasedAgreementApproveView,
+    OrganizationBasedAgreementFormView,
+    OrganizationBasedAgreementSignView,
 )
 from vitrina.orgs.views import (
     OrganizationDetailView,
@@ -138,18 +141,28 @@ urlpatterns = [
     ),
     path(
         "orgs/<int:pk>/agreements/",
-        OrganizationAgreementListView.as_view(),
+        OrganizationBasedAgreementListView.as_view(),
         name="organization-agreement-list",
     ),
     path(
         "orgs/<int:pk>/agreements/<uuid:agreement_id>/",
-        OrganizationAgreementDetailView.as_view(),
+        OrganizationBasedAgreementDetailView.as_view(),
         name="organization-agreement-detail",
     ),
     path(
-        "orgs/<int:pk>/agreements/<uuid:agreement_id>/",
-        OrganizationAgreementDetailView.as_view(),
-        name="organization-agreements-detail",
+        "orgs/<int:pk>/agreements/<uuid:agreement_id>/approve",
+        OrganizationBasedAgreementApproveView.as_view(),
+        name="organization-agreement-approve",
+    ),
+    path(
+        "orgs/<int:pk>/agreements/<uuid:agreement_id>/form",
+        OrganizationBasedAgreementFormView.as_view(),
+        name="organization-agreement-form",
+    ),
+    path(
+        "orgs/<int:pk>/agreements/<uuid:agreement_id>/sign",
+        OrganizationBasedAgreementSignView.as_view(),
+        name="organization-agreement-sign",
     ),
     path(
         "register/<token>/",