+++ CHANGES.rst
@@ -21,6 +21,7 @@ https://github.com/atviriduomenys/katalogas/issues/2040
 
 - Part 1-2: Organization-based detail & list agreement pages (previously only available under Projects).
 - Part 3: Refactor agreements to be easily extendable to other pages.
+- Part 4: Add separate agreement-negotiation views that are organization-based (reachable from organization tabs).
 
 https://github.com/atviriduomenys/katalogas/issues/2075